+++ weldx/transformations/local_cs.py
@@ -23,13 +23,16 @@ from weldx.transformations.types import (
 from weldx.transformations.util import normalize
 from weldx.types import types_time_like, types_timestamp_like
 
+from ..time import TimeDependent
+
+
 if TYPE_CHECKING:  # pragma: no cover
     import matplotlib.axes
 
 __all__ = ("LocalCoordinateSystem",)
 
 
-class LocalCoordinateSystem:
+class LocalCoordinateSystem(TimeDependent):
     """Defines a local cartesian coordinate system in 3d.
 
     Notes